LF-3985 Refactor into two-table solution
